requirements.txt
apache-airflow
python-dotenv
requests
azure-core
azure-storage-blob
apache-airflow-providers-common-compat
apache-airflow-providers-apache-spark
apache-airflow
pyspark
pyspark-stubs